Restructure hooks directory
