+++ components/address/AddressHeader.tsx
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import ChainSelect from 'components/common/ChainSelect';
+import { useNameLookup } from 'lib/hooks/ethereum/useNameLookup';
 import { useAddressPageContext } from 'lib/hooks/page-context/AddressPageContext';
 import { useMounted } from 'lib/hooks/useMounted';
-import { useNameLookup } from 'lib/hooks/useNameLookup';
 import useTranslation from 'next-translate/useTranslation';
 import { useProvider } from 'wagmi';
 import AddressDisplay from './AddressDisplay';
