+++ components/allowances/table/cells/AllowanceCell.tsx
@@ -1,8 +1,8 @@
 import { PencilIcon } from '@heroicons/react/24/outline';
 import ControlsWrapper from 'components/allowances/controls/ControlsWrapper';
 import Button from 'components/common/Button';
+import { useRevoke } from 'lib/hooks/ethereum/useRevoke';
 import { useAddressPageContext } from 'lib/hooks/page-context/AddressPageContext';
-import { useRevoke } from 'lib/hooks/useRevoke';
 import type { AllowanceData } from 'lib/interfaces';
 import { getAllowanceI18nValues } from 'lib/utils/allowances';
 import { classNames } from 'lib/utils/styles';
