+++ components/allowances/table/cells/SpenderCell.tsx
@@ -2,8 +2,8 @@ import { useQuery } from '@tanstack/react-query';
 import Href from 'components/common/Href';
 import Spinner from 'components/common/Spinner';
 import WithHoverTooltip from 'components/common/WithHoverTooltip';
+import { useOpenSeaProxyAddress } from 'lib/hooks/ethereum/useOpenSeaProxyAddress';
 import { useAddressPageContext } from 'lib/hooks/page-context/AddressPageContext';
-import { useOpenSeaProxyAddress } from 'lib/hooks/useOpenSeaProxyAddress';
 import type { AllowanceData } from 'lib/interfaces';
 import { shortenString } from 'lib/utils';
 import { getChainExplorerUrl } from 'lib/utils/chains';
